wallet/requirements.txt
Flask==3.1.0
requests==2.32.3
pika==1.3.2
python-dotenv==1.0.1
supabase==2.13.0
Flask-Cors==5.0.0